RunFgseaOnRankList - wrapper around fast gene set enrichment analysis with the fgsea R package https://bioconductor.org/packages/release/bioc/html/fgsea.html

RunFgseaOnRankList - wrapper around fast gene set enrichment analysis with the fgsea R package https://bioconductor.org/packages/release/bioc/html/fgsea.html

RunFgseaOnRankList(
  rank.list.celltype,
  pathways,
  maxSize = 500,
  minSize = 9,
  nperm = 250000,
  positive.enrich.only = FALSE
)

Arguments

rank.list.celltype
results returned by GetRankResultsRaw or GetRankResults

pathways
modules / gene sets as a named list each a single vector of unique gene IDS

maxSize
see fgsea package

minSize
see fgsea package

nperm
recommended to keep set at 25000 based on optomization and p value stabilization

positive.enrich.only
include negative enrichments in results? TRUE/FALSE

Value

results from fgsea package indexed by celltype

Examples

# usage:
t1hvl_rank = GetRankResultsRaw(limma.fit.object.list  = ebf, coefficient.number = 1, contrast.name = "time_1_highvslow")
gsea = RunFgseaOnRankList(rank.list.celltype = t1hvl_rank, )