TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast

TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast

TopGenesTidySampleExprs(
  av.exprs.list,
  result.list,
  P.Value.filter,
  logFC.filter,
  top_n_genes = NULL
)

Arguments

av.exprs.list
object returned by `PseudobulkList` summed or average counts

result.list
object returned by object returned by `GetContrastResultsRaw()` or `GetContrastResults()`

P.Value.filter
filter results

logFC.filter
filter results

top_n_genes
instead of stat thresholds above, get expression of n genes specified by this param, subsets each celltype by the top n genes ranked by t statistic.

Value

a list of tidy dataframes by celltype