WeightedCellModuleScore - calculate the average or weighted aveage +/- scaling return dataframe of cells by average module score

WeightedCellModuleScore - calculate the average or weighted aveage +/- scaling return dataframe of cells by average module score

WeightedCellModuleScore(
  gene_matrix = NULL,
  module_list,
  threshold = 0,
  cellwise_scaling = FALSE,
  return_weighted = FALSE
)

Arguments

gene_matrix

- normalized genes (rows) by cells (columns)

module_list

- names list of gene modules - each element a vector of gene names.

threshold

- at least this fraction of genes in the signature must be < 0 across all cells or else a score is not calculated

cellwise_scaling

- T/F scale across cells ?

return_weighted

- T/F weight the averae by multiplying by gene representation?

Value

dataframe of cells barcodes (rownames) by gene module scores (columns)

Examples

if (FALSE) {
}