dreamMixedModel - run dream mixed model

dreamMixedModel - run dream mixed model

dreamMixedModel(
  dge_lists,
  apriori_contrasts = FALSE,
  sample_column,
  contrast_matrix = NULL,
  design_matrix,
  fixed_effects,
  cell_metadata,
  lme4_formula = "~ 0 + cohort_timepoint + (1|sampleid)",
  ncores = 4,
  ...
)

Arguments

dge_lists

list of dgelists created with NormalizePseudobulk

apriori_contrasts

one of TRUE or FALSE, whether to fit a priori contrasts

sample_column

quoted character e.g. "sample" the subject level sample variable should have multiple timepoints subjectID_timepoint i.e. s1_0, s1_1

contrast_matrix

contrast matrix created with make.contrasts

design_matrix

design matrix created with BulkDesignMatrix

fixed_effects

a vector of covariates that are columns in metadata e.g. a vector of 'age' 'gender'

cell_metadata

metadata cells-rows variables-columns i.e. ColData or seuratmeta.data

lme4_formula

symbolic model formula the default is '~ 0 + cohort_timepoint + (1|sampleid)'

ncores

number of cores for doParallel

plotsavepath

a path to save created plot of contrasts

Value

list of model fits indexed by celltype

Examples

if (FALSE) {
c_mat = makeContrasts(
foldchange_difference = (group_timepoint1_1 - group_timepoint1_0) - (group_timepoint0_1 - group_timepoint0_0),
time1_foldchange = (group_timepoint1_1 + group_timepoint0_1) / 2  - (group_timepoint1_0 + group_timepoint0_0) / 2,
baseline_groups = (group_timepoint1_0 - group_timepoint0_0),
levels = colnames(designmat)
)
# fit mixed model for the multi timepoint contrasts
fit = scglmmr::dreamMixedModel(dge_lists = dge,
apriori_contrasts = TRUE,
sample_column = 'sample',
cell_metadata = meta,
contrast_matrix = c_mat,
design_matrix = designmat,
lme4_formula =  '~ 0 + age + gender + cohort_timepoint + (1|sampleid)',
fixed_effects = c('age', 'gender', 'cohort_timepoint'),
plotsavepath = figpath,
ncores = 4)
}
# run dream mixed model