SubjectCelltypeTable - make a QC contingency table of cells by samples; this function sets up the workflow and is run prior to running PseudobulkList

SubjectCelltypeTable - make a QC contingency table of cells by samples; this function sets up the workflow and is run prior to running PseudobulkList

SubjectCelltypeTable(metadata, celltype_column, sample_column)

Arguments

metadata
metadata cells-rows variables-columns i.e. ColData or seurat meta.data

celltype_column
quoted character e.g. "celltype" - the celltypes / clusters for which to create bulk libraries

sample_column
quoted character e.g. "sample" the subject level sample variable - if multiple timepoints should be subjectID_timepoint i.e. s1_0, s1_1

Value

a list with index 1 = contingency table of cells by sample, 2 = low representation celltypes < 10 cels, 3 celltypes to remove with 0 cells for some samples.

Examples

if (FALSE) {
datapath = "mypath/"

# load seurat or sce object etc.
s = readRDS("path/seuratobject.rds")

# define counts and metadata and subset to cells above rm seurat object from workspace
# meta = s@meta.data
# umi = s@assays$RNA@counts
# rm(s); gc()

# QC contingency of cells by subject for each celltype
tab = scglmmr::SubjectCelltypeTable(metadata = meta, celltype_column = "lineage", sample_column = "sample")
tab$celltypes_remove; tab$`low representation celltypes`; tab$table

# remove cells prior to pseudobulk analysis
meta = meta[!meta$celltype_label_3 %in% tab$celltypes_remove, ]

# subset data
umi = umi[ ,rownames(meta)]
# proceed to PseudobulkList
}