dreamMixedModel - run dream mixed model note due to this issue must require variancepartition https://github.com/GabrielHoffman/variancePartition/issues/17

dreamMixedModel - run dream mixed model note due to this issue must require variancepartition https://github.com/GabrielHoffman/variancePartition/issues/17

dreamMixedModel(
  dge_lists,
  apriori_contrasts = FALSE,
  sample_column,
  contrast_matrix = NULL,
  design_matrix,
  fixed_effects,
  cell_metadata,
  lme4_formula = "~ 0 + cohort_timepoint + (1|sampleid)",
  plotsavepath,
  ncores = 4
)

Arguments

dge_lists
list of dgelists created with NormalizePseudobulk

apriori_contrasts
one of TRUE or FALSE, whether to fit a priori contrasts

sample_column
quoted character e.g. "sample" the subject level sample variable should have multiple timepoints subjectID_timepoint i.e. s1_0, s1_1

contrast_matrix
contrast matrix created with make.contrasts

design_matrix
design matrix created with BulkDesignMatrix

fixed_effects
a vector of covariates that are columns in metadata e.g. a vector of 'age' 'gender'

cell_metadata
metadata cells-rows variables-columns i.e. ColData or seuratmeta.data

lme4_formula
symbolic model formula the default is '~ 0 + cohort_timepoint + (1|sampleid)'

plotsavepath
a path to save created plot of contrasts

ncores
number of cores for doParallel

Value

list of model fits indexed by celltype

Examples

# run dream mixed model